Query the installed driver version correctly
Most recent driver versions usually have triple digits, but the current
grep command only accounts for the driver being two digits long, so
setting nvidia_driver_version to "512.125.06" would always attempt to
install the driver again because the last digit is truncated.

With the nvidia-smi command it is easier to fetch the version.